cookie-store/unprefixed_get_all_path_and_domain.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta name="timeout" content="long">
<title>Set unprefixed cookies and getAll (path and domain shadowing)</title>
<meta name="help"
      href="https://github.com/WICG/cookie-store/blob/gh-pages/explainer.md">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/testharness-helpers.js"></script>
<script>
'use strict';

const kMyPath = location.pathname.replace(/[^\/]+$/, '');
const kMyDomain = document.domain;

// Cookie names: name1, name2, name3
// Cookie values: value1, value2 + suffix codes, value3
// Suffix codes for value2:
// - _implicit_domain_<domain> - Implicit domain=<domain>
//    NOTE: not actually the same as explicit except in IE/Edge
// - _explicit_domain_<domain> - Explicit domain=<domain>
// - _implicit_path_<path> - Implicit path=<path>
// - _explicit_path_<path> - Explicit path=<path>
const kImplicitDomain = '_implicit_domain_' + kMyDomain;
const kExplicitDomain = '_explicit_domain_' + kMyDomain;
const kImplicitRoot = '_implicit_path_/';
const kExplicitRoot = '_explicit_path_/';
const kExplicitMyPath = '_explicit_path_' + kMyPath;

promise_test(async testCase => {
  await cookieStore.set('name1', 'value1' + kImplicitDomain + kImplicitRoot);
  await cookieStore.set('name2', 'value2' + kImplicitDomain + kImplicitRoot);
  await cookieStore.set('name3', 'value3' + kImplicitDomain + kImplicitRoot);
  await cookieStore.set('name2', 'value2' + kImplicitDomain + kExplicitMyPath, {
    path: kMyPath,
  });
  await cookieStore.set(
      'name2',
      'value2' + kExplicitDomain + kImplicitRoot,
      {
        domain: kMyDomain,
      });
  await cookieStore.set(
      'name2',
      'value2' + kExplicitDomain + kExplicitMyPath,
      {
        domain: kMyDomain,
        path: kMyPath,
      });
  try {
    assert_equals(
        cookieString(await cookieStore.getAll()),
        'name2=value2' + kExplicitDomain + kImplicitRoot + '; ' +
          'name2=value2' + kExplicitDomain + kExplicitMyPath + '; ' +
          'name1=value1' + kImplicitDomain + kImplicitRoot + '; ' +
          'name2=value2' + kImplicitDomain + kImplicitRoot + '; ' +
          'name3=value3' + kImplicitDomain + kImplicitRoot);
  } finally {
    await cookieStore.delete('name1');
    await cookieStore.delete('name2');
    await cookieStore.delete('name3');
    await cookieStore.delete('name2', {path: kMyPath});
    await cookieStore.delete('name2', {domain: kMyDomain});
    await cookieStore.delete('name2', {domain: kMyDomain, path: kMyPath});
  }
}, 'getAll item order with domain and implicit path shadowing and unshadowed cookies');

promise_test(async testCase => {
  await cookieStore.set('name1', 'value1' + kImplicitDomain + kExplicitRoot, {
    path: '/',
  });
  await cookieStore.set('name2', 'value2' + kImplicitDomain + kExplicitRoot, {
    path: '/',
  });
  await cookieStore.set('name3', 'value3' + kImplicitDomain + kExplicitRoot, {
    path: '/',
  });
  await cookieStore.set('name2', 'value2' + kImplicitDomain + kExplicitMyPath, {
    path: kMyPath,
  });
  await cookieStore.set('name2', 'value2' + kExplicitDomain + kExplicitRoot, {
    path: '/',
    domain: kMyDomain,
  });
  await cookieStore.set('name2', 'value2' + kExplicitDomain + kExplicitMyPath, {
    domain: kMyDomain,
    path: kMyPath,
  });
  try {
    assert_equals(
        cookieString(await cookieStore.getAll()),
        'name2=value2' + kExplicitDomain + kExplicitRoot + '; ' +
          'name2=value2' + kExplicitDomain + kExplicitMyPath + '; ' +
          'name1=value1' + kImplicitDomain + kExplicitRoot + '; ' +
          'name2=value2' + kImplicitDomain + kExplicitRoot + '; ' +
          'name3=value3' + kImplicitDomain + kExplicitRoot);
  } finally {
    await cookieStore.delete('name1', {path: '/'});
    await cookieStore.delete('name2', {path: '/'});
    await cookieStore.delete('name3', {path: '/'});
    await cookieStore.delete('name2', {path: kMyPath});
    await cookieStore.delete('name2', {path: '/', domain: kMyDomain});
    await cookieStore.delete('name2', {domain: kMyDomain, path: kMyPath});
  }
}, 'getAll item order with domain and explicit path shadowing and unshadowed cookies');

</script>
